# lib/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: nilsbrugger.nb.nb.nb@gmail.com
    name: Nils
  description: The time planer API
  license:
    name: Apache 2.0
  title: Time Planer
  version: 2.2.1-alpha
servers:
- url: http://eds.logfro.de/time-planer
- url: http://localhost:8080/time-planer
...
paths:
  /auth:
    get:
      description: Here you authenticate as the user and get his API key
      operationId: login
      parameters:
      - description: The Email Adress of the User
        explode: true
        in: query
        name: email
        required: true
        schema:
          format: email
          type: string
        style: form
      - description: The Uses password
        explode: true
        in: query
        name: password
        required: true
        schema:
          format: password
          type: string
        style: form
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
          description: login successfull
        203:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: no password
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request does not contains a passwort for the user
        204:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: api key wrong
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: No email address
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: params not matching pattern
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            Some of the parameters did not matched the pattern they should.
            For example an email adress needs to match the pattern
            `{some chars}@{provider}.{top level domain}`
        401:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: wrong password
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: wrong password
        404:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: user not found
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: There is no user with such an email address. Maybe he deleted his account
      summary: Login
      tags:
      - Authentication
      x-accepts: application/json
      x-tags:
      - tag: Authentication
    put:
      description: |-
        Registers a new user.
        ### Requirements
        * Name
        * Email
        * Password
      operationId: registrate
      requestBody:
        $ref: '#/components/requestBodies/RegistrationRequest'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
          description: regisration worked
        400:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: params not matching pattern
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: |-
            Some of the parameters did not matched the pattern they should.
            For example an email adress needs to match the pattern
            `{some chars}@{provider}.{top level domain}`
        404:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Params missing
        409:
          content:
            application/json:
              examples:
                response:
                  value:
                    message: User allready exists
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: User allready exists
      summary: Registration
      tags:
      - Authentication
      x-accepts: application/json
      x-tags:
      - tag: Authentication
      x-contentType: application/json
...
components:
...
  requestBodies:
...
    RegistrationRequest:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RegistrationRequest'
...
  schemas:
...
    RegistrationRequest:
      additionalProperties: false
      description: |-
        ### Requirements
        * Name
        * Email
        * Password
      example:
        password: password
        name: name
        email: nbrugger@studnet.tgm.ac.at
      properties:
        name:
          description: |-
            The name of an user.

            unique
          maxLength: 20
          minLength: 3
          type: string
        email:
          example: nbrugger@studnet.tgm.ac.at
          format: email
          type: string
        password:
          description: |-
            # Password
              * Min length: 8
              * Max Length : 40
              * At least one small capitalized char
              * One upper case char
              * pattern : ^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).{8,40}$
          format: password
          maxLength: 40
          minLength: 8
          pattern: ^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).{8,40}$
          title: Password
          type: string
      required:
      - email
      - name
      - password
      type: object
...
    ErrorResponse:
      description: Detailed information about an Error
      properties:
        message:
          description: A short description only a few words, what went wrong
          enum:
          - api key missing
          - api key wrong
          - group not found
          - group not owned
          - no password
          - no permission
          - user not found
          - params not matching pattern
          - No email address
          - wrong password
          - User allready exists
          - task_not_found
          type: string
      required:
      - message
      type: object
    SuccessResponse:
      description: The Request was successfull handled
      example:
        message: registration worked
      properties:
        message:
          description: A short mesage what went good
          enum:
          - registration worked
          - all tasks deleted
          type: string
      required:
      - message
      type: object
    inline_response_200:
      example:
        user_key: user_key
      properties:
        user_key:
          description: |-
            Every user has an uniqe API key.
            Whith this Key you can access all datas and actions
          format: password
          maxLength: 1000
          minLength: 500
          pattern: ^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z]).{500,1000}$
          type: string
  securitySchemes:
    User_Key:
      in: header
      name: user_key
      type: apiKey